Move some to its own file that imports array, fix up an issue related to how I was templatizing the iterators.
